requirements.txt
PyQt5
pyqtgraph
pyserial
pyqtgraph
